:recycle: Move DSP UI structs to view modules
