+++ src/dsp/mod.rs
@@ -1,9 +1,6 @@
 //! Digital signal processing functions and user interface logic.
 
-pub mod base;
 pub mod buffer;
-pub mod normalize;
+pub mod filters;
 
-pub use crate::dsp::base::{Filter, Knob};
 pub use crate::dsp::buffer::Samples;
-pub use crate::dsp::normalize::Normalize;
